+++ lib/src/ast/sass/expression/number.dart
@@ -6,6 +6,7 @@ import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 
 import '../../../visitor/interface/expression.dart';
+import '../../../value/number.dart';
 import '../expression.dart';
 
 /// A number literal.
@@ -14,17 +15,18 @@ import '../expression.dart';
 @sealed
 class NumberExpression implements Expression {
   /// The numeric value.
-  final num value;
+  final double value;
 
   /// The number's unit, or `null`.
   final String? unit;
 
   final FileSpan span;
 
-  NumberExpression(this.value, this.span, {this.unit});
+  NumberExpression(num value, this.span, {this.unit})
+      : value = value.toDouble();
 
   T accept<T>(ExpressionVisitor<T> visitor) =>
       visitor.visitNumberExpression(this);
 
-  String toString() => "$value${unit ?? ''}";
+  String toString() => SassNumber(value, unit).toString();
 }