+++ pkg/sass_api/CHANGELOG.md
@@ -5,6 +5,10 @@
 * Add an optional `argumentName` parameter to `SassScriptException()` to make it
   easier to throw exceptions associated with particular argument names.
 
+* Most APIs that previously returned `num` now return `double`. All APIs
+  continue to _accept_ `num`, although in Dart 2.0.0 most of these APIs will be
+  changed to accept only `double`.
+
 ## 3.0.4
 
 * `UnaryOperationExpression`s with operator `not` now include a correct span,